% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunctions.R
\name{fitModel}
\alias{fitModel}
\title{Fit a Gaussian mixture model}
\usage{
fitModel(
  data,
  trait.colums,
  factor.column,
  n.pop,
  n.burnin = 5000,
  n.sampling = 10000,
  n.thin = 5,
  n.chains = 5
)
}
\arguments{
\item{data}{data frame with the data}

\item{factor.column}{name of the column that contains the factor connecting replicates to lines}

\item{n.pop}{number of populations to fit, must be two ro greater}

\item{n.burnin}{number of iterations of butnin (adaptation)}

\item{n.sampling}{number of sampling steps}

\item{n.thin}{thinning number (if, e.g., set to five then every fifth chain sample is saved)}

\item{n.chains}{number of chains}

\item{trait.columns}{list of columns in \code{data} that contain trait values}
}
\value{
S3 object of class \code{mugamix} that contains matrix of parameter chains (named \code{parChains}, each chain a column), a matrix of population assignments (named \code{popChains}), and a matrix of population numbers (named \code{nPopsChain})
}
\description{
Fits a mixture model to the provided data, taking into account replication structure. Takes data on multiple traits and generates samples from posterior distributions of parameters, as well as probabilities that a line belongs to a given population. The fit is performed using a No-U-Turn Sampler (NUTS). The recommended burn-in, sampling, number of chains, and thinning are set as defaults.
}
\note{
Currently exactly one level of replication is supported, with no missing data.
}
