% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunctions.R
\name{plot.vbmugamix}
\alias{plot.vbmugamix}
\title{Plot variational Bayes population assignments}
\usage{
\method{plot}{vbmugamix}(obj)
}
\arguments{
\item{obj}{a \code{vbmugamix} object generated by the \code{quickFitModel} function.}
}
\value{
a plot object (either \code{ggplot} or \code{barplot})
}
\description{
Plot method for MuGaMix objects (generated by \code{quickFitModel()}). Plots population assignment probabilities for each line. Uses \code{data.table} and \code{ggplot2} if available.
If the data set contains many lines, their names may not show up well. In such cases, the user can further modify plot parameters by using the plot objects returned by the function.
}
