% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSampler}
\alias{runSampler}
\title{Run the sampler}
\usage{
runSampler(yVec, lnFac, Npop, Nadapt, Nsamp, Nthin, Nchains)
}
\arguments{
\item{yVec}{vectorized data matrix}

\item{lnFac}{factor relating data points to lines}

\item{Npop}{number of populations}

\item{Nadapt}{number of adaptation (burn-in) steps}

\item{Nsamp}{number of sampling steps}

\item{Nthin}{thinning number}
}
\description{
Runs the sampler on the data assuming no fixed effects or missing trait data and one replication level.
}
