% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSamplerMiss}
\alias{runSamplerMiss}
\title{Run the sampler with missing data}
\usage{
runSamplerMiss(yVec, lnFac, missIDs, Npop, Nadapt, Nsamp, Nthin, Nchains)
}
\arguments{
\item{yVec}{vectorized data matrix}

\item{lnFac}{factor relating data points to lines}

\item{missIDs}{vectorized matrix (same dimensions as data) with 1 where a data point is missing and 0 otherwise}

\item{Npop}{number of populations}

\item{Nadapt}{number of adaptation (burn-in) steps}

\item{Nsamp}{number of sampling steps}

\item{Nthin}{thinning number}
}
\description{
Runs the sampler on the data assuming no fixed effects, but allowing for missing phenotype data, and one replication level.
The missingness indicator should have 1 for missing data points and 0 otherwise, however any non-0 value is treated as 1.
}
