% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vbFit}
\alias{vbFit}
\title{Variational Bayes model fit}
\usage{
vbFit(yVec, d, nGroups, alphaPr, sigSqPr, covRatio, nReps)
}
\arguments{
\item{yVec}{vectorized data matrix}

\item{d}{number of traits}

\item{nGroups}{number of groups}

\item{alphaPr}{prior group size}

\item{sigSqPr}{prior variance}

\item{covRatio}{population to error covariance ratio}

\item{nReps}{number of model fit attempts before picking the best fit}
}
\value{
list containing group means (\code{groupMeans}), covariances (\code{covariances}), effective group sizes (\code{effNm}), group assignment probabilities (\code{p}), and the deviance information criterion (DIC, \code{DIC}).
}
\description{
Fits a Gaussian mixture model using variational Bayes. Assumes no missing data.
}
\keyword{internal}
