% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunctions.R
\name{plot.mugamix}
\alias{plot.mugamix}
\title{Plot population assignments}
\usage{
\method{plot}{mugamix}(obj)
}
\arguments{
\item{obj}{a \code{mugamix} object generated by the \code{fitModel} function.}
}
\description{
Plots population assignment probabilities per line. Uses \code{data.table} and \code{ggplot2} if available.
}
